fix compilation warning
git-svn-id: svn+ssh://svn.gna.org/svn/xenomai/trunk@508 c6d672ea-8702-0410-b560-f74c916a59fe

+++ ksrc/skins/native/heap.c
@@ -63,7 +63,7 @@ static int __heap_read_proc (char *page,
     p += sprintf(p,"type=%s:size=%lu:used=%lu\n",
 		 (heap->mode & H_SHARED) == H_SHARED ? "shared" :
 		 (heap->mode & H_MAPPABLE) ? "mappable" : "kernel",
-		 heap->csize,
+		 (u_long) heap->csize,
 		 xnheap_used_mem(&heap->heap_base));
 
     xnlock_get_irqsave(&nklock,s);
